requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.29.0
flask==2.3.3
werkzeug==2.3.7
python-dotenv==1.1.1
pinecone-client==3.2.2
psycopg2-binary==2.9.7
sqlalchemy==2.0.23
google-generativeai==0.3.2
python-multipart==0.0.6
requests==2.31.0
PyPDF2==3.0.1
python-docx==1.1.0
pandas==2.1.4
numpy==1.26.4
pydantic==2.5.2
httpx==0.28.1
aiofiles==23.2.0
scikit-learn==1.5.1
